Added error message if not all control paths of a function return a value

+++ src/Ast.h
@@ -117,7 +117,7 @@ struct Ir {
 		} location;
 	};
 
-	u64 flags = 0;
+	u8 flags = 0;
 	IrOp op;
 };
